Refactor code to BaseReport and fix broken tests
