+++ lib/open_food_network/bulk_coop_report.rb
@@ -1,9 +1,9 @@
-require 'open_food_network/reports/report'
+require 'open_food_network/reports/base_report'
 require 'open_food_network/reports/bulk_coop_supplier_report'
 require 'open_food_network/reports/bulk_coop_allocation_report'
 
 module OpenFoodNetwork
-  class BulkCoopReport < Reports::Report
+  class BulkCoopReport < Reports::BaseReport
 
     def initialize(user, params = {})
       @params = params
