+++ spec/lib/open_food_network/order_and_distributor_report_spec.rb
@@ -1,6 +1,5 @@
 require 'spec_helper'
 
-
 module OpenFoodNetwork
   describe OrderAndDistributorReport do
 
@@ -96,24 +95,8 @@ module OpenFoodNetwork
       end
 
       context "Old version" do
-        before(:each) do
-          #normal completed order
-          @bill_address = create(:address)
-          @distributor_address = create(:address, :address1 => "distributor address", :city => 'The Shire', :zipcode => "1234")
-          @distributor = create(:distributor_enterprise, :address => @distributor_address)
-          product = create(:product)
-          product_distribution = create(:product_distribution, :product => product, :distributor => @distributor)
-          @shipping_instructions = "pick up on thursday please!"
-          @order = create(:order, :distributor => @distributor, :bill_address => @bill_address, :special_instructions => @shipping_instructions)
-          @payment_method = create(:payment_method, :distributors => [@distributor])
-          payment = create(:payment, :payment_method => @payment_method, :order => @order )
-          @order.payments << payment
-          @line_item = create(:line_item, :product => product, :order => @order)
-          @order.line_items << @line_item
-        end
-
         it "should return a header row describing the report" do
-          subject = OrderAndDistributorReport.new user, {}, [@order]
+          subject = OrderAndDistributorReport.new user
 
           header = subject.header
           header.should == ["Order date", "Order Id",
@@ -123,16 +106,33 @@ module OpenFoodNetwork
             "Distributor", "Distributor address", "Distributor city", "Distributor postcode", "Shipping instructions"]
         end
 
-        it "should denormalise order and distributor details for display as csv" do
-          subject = OrderAndDistributorReport.new user, {}, [@order]
+        context "table results" do
+          before(:each) do
+            #normal completed order
+            @bill_address = create(:address)
+            @distributor_address = create(:address, :address1 => "distributor address", :city => 'The Shire', :zipcode => "1234")
+            @distributor = create(:distributor_enterprise, :address => @distributor_address)
+            product = create(:product)
+            product_distribution = create(:product_distribution, :product => product, :distributor => @distributor)
+            @shipping_instructions = "pick up on thursday please!"
+            @order = create(:order, completed_at: 1.day.ago, :distributor => @distributor, :bill_address => @bill_address, :special_instructions => @shipping_instructions)
+            @payment_method = create(:payment_method, :distributors => [@distributor])
+            payment = create(:payment, :payment_method => @payment_method, :order => @order )
+            @order.payments << payment
+            @line_item = create(:line_item, :product => product, :order => @order)
+            @order.line_items << @line_item
+          end
 
-          table = subject.table
+          it "should denormalise order and distributor details for display as csv" do
+            subject = OrderAndDistributorReport.new user
 
-          table[0].should == [@order.created_at, @order.id,
-            @bill_address.full_name, @order.email, @bill_address.phone, @bill_address.city,
-            @line_item.product.sku, @line_item.product.name, @line_item.options_text, @line_item.quantity, @line_item.max_quantity, @line_item.price * @line_item.quantity, @line_item.distribution_fee,
-            @payment_method.name,
-            @distributor.name, @distributor.address.address1, @distributor.address.city, @distributor.address.zipcode, @shipping_instructions ]
+            table = subject.table
+            table[0].should == [@order.created_at, @order.id,
+              @bill_address.full_name, @order.email, @bill_address.phone, @bill_address.city,
+              @line_item.product.sku, @line_item.product.name, @line_item.options_text, @line_item.quantity, @line_item.max_quantity, @line_item.price * @line_item.quantity, @line_item.distribution_fee,
+              @payment_method.name,
+              @distributor.name, @distributor.address.address1, @distributor.address.city, @distributor.address.zipcode, @shipping_instructions ]
+          end
         end
       end
     end
